Separate the GDB RSP client from the rest of TA-VM
In the original TAJ proof-of-concept, as disclosed in Brescia, the GDB
client assumed the then-current implementation of PDL modeling.
Today's family of implementations is a matrix of many dimensions.
For one, on the prerequisites side, there is the Petrich1/Petrich2
difference over what a "machine integer" is.  For another, there is
ArchC vs Sail, therefore GDB now needs an abstract notion of what a PDL
is.  On the consumer side, we have Modtalk vs Bee vs OpenSmalltalk VM,
as well as different Smalltalk dialects.  This branch begins to untangle
this spaghetti of dependencies.